backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
flask-migrate
flask-jwt-extended
flask-bcrypt
gunicorn
google-generativeai
